requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.3
attrs==23.2.0
charset-normalizer==3.3.2
dnspython==2.5.0
frozenlist==1.4.1
idna==3.6
install==1.3.5
multidict==6.0.5
pip==23.0.1
py-cord==2.4.1
python-dotenv==1.0.1
setuptools==66.1.1
yarl==1.9.4
